Allow GetFlags to return nil.
This allows individual subcommands to indicate that they would like to
take full control of the 'args' parsing. This is useful for when 'flag'
is insufficiently groovy, or when writing wrapper programs where some
or even none of the flags need to be interpreted by the wrapper itself
(e.g. merely passed through to the target program).

Without this, invocations such as `prog subcommand -flag` will fail
because `-flag` would be interpreted as an 'unknown flag'.